Fix my own PR feedback
